Refactor profiling and other functions into utils submoudle
