+++ src/gains/Analysis/analyse_spin_up.py
@@ -1,6 +1,5 @@
 """Contains functions to produce plots in scripts/plot_spin_up.py."""
 
-import re
 from pathlib import Path
 
 import h5py
@@ -9,7 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import scipy.interpolate as inp
 
-from gains.utils import get_arg_of_nearest
+from gains.utils.misc import extract_numerical_suffix, get_arg_of_nearest
 
 
 class LabeledCoordinate:
@@ -24,22 +23,6 @@ class LabeledCoordinate:
         """
         self.coord = coord
         self.label = label
-
-
-def extract_numerical_suffix(path: Path) -> int | float:
-    """
-    Extract an integer at the end of a filename.
-
-    Takes a path to a file saved in the form /output_dir/file_name[num].extension
-    and return num. Files that don't fit this format will be assigned inf, so
-    placed at the end of a list when sorting.
-
-    :param path: path to the output file, in the form
-    /output_dir/file_name[num].extension.
-    :returns suffix: Integer at the end of the file name.
-    """
-    match = re.search(r"(\d+)$", path.stem)
-    return int(match.group(1)) if match else float("inf")
 
 
 def my_interp2d(f: np.ndarray, rad: np.ndarray, radnew: np.ndarray) -> np.ndarray:
